requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Training framework
lightning==2.4.0

# Data & augmentation
albumentations==1.4.11
opencv-python==4.10.0.84
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
pandas>=2.0.0
Pillow>=10.0.0

# Segmentation models
segmentation-models-pytorch>=0.3.3

# XAI
captum==0.7.0
ttach>=0.0.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml==6.0.2
rich==13.9.2

# App
streamlit==1.38.0

# Development
ruff>=0.1.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
